improves dev env up & teardown

+++ down.sh
@@ -4,5 +4,8 @@ docker stop -t=0 `docker ps -aq`
 echo "Removing stale containers..."
 docker rm `docker ps -aq`
 
-echo "Removing untagged images..."
-docker rmi $(docker images -a | grep "^<none>" | awk '{print $3}')
+if [ "$1" = "clean" ]
+then
+    echo "Removing untagged images..."
+    docker rmi $(docker images -a | grep "^<none>" | awk '{print $3}')
+fi
